+++ bee-test/src/rand/option.rs
@@ -5,9 +5,5 @@ use crate::rand::bool::rand_bool;
 
 /// Generates a random generic [`Option`].
 pub fn rand_option<T>(inner: T) -> Option<T> {
-    if rand_bool() {
-        Some(inner)
-    } else {
-        None
-    }
+    if rand_bool() { Some(inner) } else { None }
 }
